+++ ubo_app/menu_app/menu_notification_handler.py
@@ -30,6 +30,7 @@ if TYPE_CHECKING:
     from collections.abc import Callable
 
     from ubo_gui.menu.menu_widget import MenuWidget
+    from ubo_gui.menu.types import PageWidget
 
 
 class NotificationReference:
@@ -173,10 +174,8 @@ class MenuNotificationHandler(UboApp):
         if notification.value.extra_information:
             text = notification.value.extra_information.text
 
-            def open_info() -> None:
-                info_application = NotificationInfo(text=text)
-
-                store.dispatch(OpenApplicationAction(application=info_application))
+            def open_info() -> PageWidget:
+                return NotificationInfo(text=text)
 
             items.append(
                 NotificationActionItem(