+++ ubo_app/store/services/docker.py
@@ -32,6 +32,7 @@ class DockerItemStatus(StrEnum):
     CREATED = auto()
     RUNNING = auto()
     ERROR = auto()
+    PROCESSING = auto()
 
 
 class DockerAction(BaseAction):